+++ CHANGELOG.md
@@ -15,8 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Update npm dependencies
 - Move browserslist config outside webpack config
 - Move Babel config outside webpack config
-- Enable awesome-typescript-loader caching
 - Babel now recompiles TypeScript’s output using browserslist data
+- Switch from awesome-typescript-loader to ts-loader
 
 ### Fixed
 - Fix `##` that should be `###` in change log